LayoutTests/imported/w3c/web-platform-tests/svg/animations/svglength-animation-invalid-value-3-expected.txt


FAIL Test SVGLength animation with invalid value: No spaces after number and unit. assert_approx_equals: expected 100 +/- 1 but got 0
